Removed responsibility for like_ship merging from OOShipClass legacy initializer.
Doing it that way was turning out to be very restricting. It becomes impossible to remove/simplify existing properties or implement obvious constraints (like only defining carrier-related stuff in carriers).

I came close to passing the like_ship’s definition dictionary alongside its OOShipClass, but at that point it would be easier to just merge the dictionaries and work from that.

All in all, it seems smarter to reuse the existing merging code from OOShipRegistry, which will give us perfect fidelity with 1.x as well as simplifying the init/convert code.