Basic implementation of new OOMaterial.
OOMaterials can be generated from material specification + render mesh combos. Currently, only a solid-colour shader is used.
To support shader loading, a general abstraction of file finding and loading has been added in the form of OoliteBase/OOFileResolving. This will need refinement to deal with OXZs later, but it's a start.