OoliteGraphics general texture swizzling support:
It is now possible to specify one to four channels in a texture specifier’s “extract” key, in any order and optionally repeating channels, in the same manner as GLSL swizzle operations. On the other hand, since extraction is now handled in the shader, “extract” will be ignored for materials with custom shaders. Using “extract” in those cases would be pointless and inefficient anyway; the only time it made sense in 1.x was for shaderless fallbacks, which are now irrelevant. The new implementation, for synthesized materials, _is_ efficient – a single sample per source texture is taken and components are swizzled out as needed.