OoliteSound: major refactoring:
* State is no longer stored in statics and singletons; instead, there’s
  an explicit OOSoundContext class, which acts as a factory for other
  sound-releated types.
* The OOSoundContext to use is selected at runtime, allowing multiple
  sound engines in a single binary.
* The public OOSoundSource class no longer assumes a channel-and-mixer
  model.

These changes prepare the ground for a future OpenAL sound implementation,
with the option of keeping the current SDL engine as a fallback and/or
adding 3D sound support to the Core Audio engine.

The SDL implementation has not yet been adapted to the new form.
Music is temporarily unsupported.